Add better extension for remove yellow message.
No disrespect to author of other extension for putting mine first. The reason is my extension is compatible with FireFox, Chrome, ... and is available in Chrome store and FireFox MDN. For first one, we have to install it manually.
Mine also uses Css instead of JavaScript, so its faster. (By the way i created it yesterday :D)